requirements.txt
numpy
pandas
seaborn
matplotlib
scikit-learn
catboost
xgboost
pymysql
python-dotenv
flask
mlflow
# -e .